C++ =/= C
