+++ Altis_Life.Altis/description.ext
@@ -28,19 +28,19 @@ SPY_Enable = 0;
 
 #include "Config_Spyglass.hpp"
 #include "CfgRemoteExec.hpp"
-#include "dialog\MasterHandler.h"
+#include "dialog\MasterHandler.hpp"
 #include "Config_Master.hpp"
 
 class RscTitles
 {
 	#include "statusBar\statusBar.hpp"
 	#include "dialog\progress.hpp"
-	#include "dialog\hud_nameTags.h"
+	#include "dialog\hud_nameTags.hpp"
 };
 
 class CfgFunctions
 {
-	#include "Functions.h"
+	#include "Functions.hpp"
 };
 
 class CfgSounds
