+++ Altis_Life.Altis/dialog/MasterHandler.hpp
@@ -17,10 +17,10 @@
 #include "clothing.hpp"
 #include "trunk.hpp"
 #include "spawnSelection.hpp"
-#include "chop_shop.h"
-#include "pInteraction.h"
-#include "deathScreen.h"
-#include "vehicleShop.h"
-#include "settings.h"
-#include "federalReserve.h"
+#include "chop_shop.hpp"
+#include "pInteraction.hpp"
+#include "deathScreen.hpp"
+#include "vehicleShop.hpp"
+#include "settings.hpp"
+#include "federalReserve.hpp"
 #include "compensate.hpp"
